Separate header for Candidate.
