+++ src/sequence_batch.h
@@ -212,7 +212,8 @@ class SequenceBatch {
   static constexpr char uint8_to_char_table_[8] = {'A', 'C', 'G', 'T',
                                                    'N', 'N', 'N', 'N'};
   void ReplaceByEffectiveRange(kstring_t &seq) {
-    if (effective_range_[0] == 0 && effective_range_[1] == -1 && effective_range_[2] == 1) {
+    if (effective_range_[0] == 0 && effective_range_[1] == -1 &&
+        effective_range_[2] == 1) {
       return;
     }
     int i, j;
@@ -225,10 +226,10 @@ class SequenceBatch {
     seq.s[i] = '\0';
     seq.l = end - start + 1;
     if (effective_range_[2] == -1) {
-      for (i = 0 ; i < (int)seq.l ; ++i) {
-        seq.s[i] = Uint8ToChar(((uint8_t)3)^(CharToUint8(seq.s[i])));
+      for (i = 0; i < (int)seq.l; ++i) {
+        seq.s[i] = Uint8ToChar(((uint8_t)3) ^ (CharToUint8(seq.s[i])));
       }
-      for (i = 0, j = seq.l - 1; i < j ; ++i, --j) {
+      for (i = 0, j = seq.l - 1; i < j; ++i, --j) {
         char tmp = seq.s[i];
         seq.s[i] = seq.s[j];
         seq.s[j] = tmp;
